% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_map_aspect_ratio}
\alias{get_map_aspect_ratio}
\title{Get map aspect ratio}
\usage{
get_map_aspect_ratio(map, crs)
}
\arguments{
\item{map}{A ggplot object.}

\item{crs}{A coordinate reference system (CRS).}
}
\value{
A single numeric aspect ratio (width/height).
}
\description{
Convenience wrapper around \code{\link[=get_map_range]{get_map_range()}} returning only the numeric
aspect ratio.
}
\keyword{internal}
