% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_map_range}
\alias{get_map_range}
\title{Get Map Range and Aspect Ratio}
\usage{
get_map_range(map, crs)
}
\arguments{
\item{map}{A ggplot object containing a map.}

\item{crs}{A coordinate reference system specification.}
}
\value{
A list with components:
\item{xrange}{The range of x-coordinates}
\item{yrange}{The range of y-coordinates}
\item{aspect_ratio}{The aspect ratio of the map}
}
\description{
Calculates the coordinate ranges and aspect ratio of a ggplot map object,
taking into account the coordinate reference system.
}
\keyword{internal}
