% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset_config.R
\name{plot_spec}
\alias{plot_spec}
\title{Create Plot Specification for Inset Maps}
\usage{
plot_spec(
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  loc_left = 0,
  loc_bottom = 0,
  width = NULL,
  height = NULL,
  no_scale = FALSE,
  main = FALSE
)
}
\arguments{
\item{xmin, xmax, ymin, ymax}{Numeric values defining the bounding box coordinates
for cropping the spatial data. If NULL, the full extent will be used.}

\item{loc_left, loc_bottom}{Numeric values between 0 and 1 specifying the
position of the inset plot within the main plot area. Default is (0, 0).}

\item{width, height}{Numeric values specifying the width and height of the
inset plot as proportions of the main plot. If both are provided, a warning
is issued as aspect ratio maintenance is prioritized.}

\item{no_scale}{Logical. If TRUE, the inset size is determined automatically
based on the spatial extent relative to the main plot. Default is FALSE.}

\item{main}{Logical. If TRUE, this specification defines the main plot.
Only one specification should have main = TRUE. Default is FALSE.}
}
\value{
A list containing the plot specification parameters.
}
\description{
Defines the spatial extent and positioning parameters for individual plots
in an inset map configuration.
}
\examples{
# Create a main plot specification
main_spec <- plot_spec(main = TRUE)

# Create an inset plot specification
inset_spec <- plot_spec(
    xmin = -120, xmax = -100, ymin = 30, ymax = 50,
    loc_left = 0.7, loc_bottom = 0.7,
    width = 0.3
)

}
