% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_map_aspect_ratio}
\alias{get_map_aspect_ratio}
\title{Get Map Aspect Ratio}
\usage{
get_map_aspect_ratio(map, crs)
}
\arguments{
\item{map}{A ggplot object containing a map.}

\item{crs}{A coordinate reference system specification.}
}
\value{
Numeric value representing the aspect ratio.
}
\description{
A convenience function that returns only the aspect ratio of a map.
}
\keyword{internal}
