% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset_config.R
\name{last_insetcfg}
\alias{last_insetcfg}
\title{Get Last Inset Configuration}
\usage{
last_insetcfg()
}
\value{
An inset configuration object of class "insetcfg", or NULL if no
configuration has been set.
}
\description{
Retrieves the most recently stored inset configuration object.
This is used internally by \code{\link[=with_inset]{with_inset()}} when no configuration is explicitly provided.
}
\examples{
\dontrun{
# After setting up a configuration
config_insetmap(some_data, specs = some_specs)

# Retrieve the configuration
cfg <- last_insetcfg()
}

}
