% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_border}
\alias{map_border}
\title{Add Border to Map Plot}
\usage{
map_border(color = "black", linewidth = 1, fill = "white", ...)
}
\arguments{
\item{color}{Character string specifying the border color. Default is "black".}

\item{linewidth}{Numeric value specifying the border line width. Default is 1.}

\item{fill}{Character string specifying the background fill color. Default is "white".}

\item{...}{Additional arguments passed to \code{\link[ggplot2:element]{ggplot2::element_rect()}}.}
}
\value{
A ggplot2 theme object that can be added to a plot.
}
\description{
Creates a theme element that adds a border around a map plot. Useful for
distinguishing inset plots from the main plot.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(mpg, wt)) +
    geom_point() +
    map_border(color = "red", linewidth = 2)

}
